split functions in seperate files
