+++ src/border_cam.c
@@ -1,0 +1,22 @@
+#include "fita.h"
+
+int border_cam(cursor_t *cam)
+{
+	int width = GET_WIDTH;
+	int height = GET_HEIGHT;
+
+	if ((cam->x < ((width / 2) + 1) && cam->y < ((height / 2) + 1)) ||
+		(cam->x < ((width / 2) + 1) && cam->y > 512 - (height / 2)) ||
+		(cam->x > 512 - (width / 2) && cam->y < ((height / 2) + 1)) ||
+		(cam->x > 512 - (width / 2) && cam->y > 512 - (height / 2)))
+		return 5;
+	if (cam->x < ((width / 2) + 1))
+		return 4;
+	if (cam->x > 512 - (width / 2))
+		return 2;
+	if (cam->y < ((height / 2) + 1))
+		return 1;
+	if (cam->y > 512 - (height / 2))
+		return 3;
+	return 0;
+}
